backend2or0/ow/requirements.txt
flask
mysql-connector-python
DateTime
requests
pyyaml